GRP-1391 - PSPNG - Failing on missing members
Changed PSPNG to just warn on when subjects can't be found in target system. The
changelog events referring to such missing subjects are marked completed, meaning
that they'll be provisioned into groups on the first full-sync after they're created
in the target system.

+++ grouper-misc/grouper-pspng/src/main/java/edu/internet2/middleware/grouper/pspng/Provisioner.java
@@ -902,8 +902,9 @@ public abstract class Provisioner
         TSUserClass tsUser = tsUserCache_shortTerm.get(subject);
         
         if ( config.needsTargetSystemUsers() && tsUser==null ) {
-          LOG.info("{}: Skipping membership for subject that doesn't already exist in target system: {}", getName(), subject.getName());
-          workItem.markAsSuccess("Skipping membership of subject '%s' that doesn't exist in target system", subject.getName());
+          LOG.warn("{}: Skipping adding membership to {} for subject that doesn't exist in target system: {}", 
+              new Object[]{getName(), grouperGroupInfo.getName(), subject.getName()});
+          workItem.markAsSuccess("Skipped: subject '%s' doesn't exist in target system", subject.getName());
           return;
         }
         
@@ -920,6 +921,12 @@ public abstract class Provisioner
         Subject subject = workItem.getSubject(this);
         TSUserClass tsUser = tsUserCache_shortTerm.get(subject);
   
+        if ( config.needsTargetSystemUsers() && tsUser==null ) {
+          LOG.warn("{}: Skipping removing membership of subject from {}: Subject doesn't already exist in target system: {}",
+              new Object[]{getName(), grouperGroupInfo.getName(), subject.getName()});
+          workItem.markAsSuccess("Skipped: subject '%s' doesn't exist in target system", subject.getName());
+          return;
+        }
         deleteMembership(grouperGroupInfo, tsGroup, subject, tsUser);
       }
       else
